Refactor to split database code from model code
